# config/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Watch & Do API",
        "description": "Watch & Do API",
        "contact": {
            "name": "Alexandre Maia",
            "url": "http://wwwalexandremaia.pt",
            "email": "info@alexandremaia.pt"
        },
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/favorites": {
            "get": {
                "tags": [ "Favorites" ],
                "description": "<h2>Get User Favorites active Videos List</h2><br><p>Role required: <strong>Register</strong></p>",
                "responses": {
                    "200": {
                        "description": "Favorite Videos List",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Favorites" },
                                "examples": {
                                    "Videos List": [
                                        [
                                            {
                                              "id": 1,
                                              "topic": "HTML",
                                              "description": "Crie a estrutura de um Site"
                                            },
                                            {
                                              "id": 3,
                                              "topic": "CSS",
                                              "description": "Altere a aparência de um Site"
                                            }
                                        ]
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/favorites/{videoId}": {
            "get": {
                "tags": [ "Favorites" ],
                "description": "<h2>Add/Remove Video from User Favorites</h2><br><p>Role required: <strong>Register</strong></p>",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "description": "Video Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Add Favorite": {
                                    "value": "Video added to Favorites"
                                },
                                "Remove Favorite": {
                                    "value": "Video removed from Favorites"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/favorites": {
            "get": {
                "tags": [ "Favorites" ],
                "description": "<h2>Get Favorites List</h2><br><p>Role required: <strong>Admin</strong></p>",
                "responses": {
                    "200": {
                        "description": "Favorites List",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Favorites" },
                                "examples": {
                                    "Favorite List": 
                                    [
                                        {
                                            "id": 2,
                                            "userId": 3,
                                            "videoId": 1,
                                            "createdAt": "2020-12-02T17:42:16.000Z",
                                            "updatedAt": "2020-12-02T17:42:16.000Z"
                                        },
                                        {
                                            "id": 4,
                                            "userId": 3,
                                            "videoId": 2,
                                            "createdAt": "2020-12-04T08:33:54.000Z",
                                            "updatedAt": "2020-12-04T08:33:54.000Z"
                                          }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "post": {
                "tags": [ "Favorites" ],
                "description": "<h2>Create new Favorite</h2><br><p>Role required: <strong>Admin</strong></p>",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Favorites" },
                            "examples": {
                                "Favorite": {
                                    "value": {
                                        "userId": 1,
                                        "videoId": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New Favorite",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Favorites" }
                            },
                            "text/plain": {
                                "schema": { "type": "string" }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": { "type": "string" },
                                "examples": {
                                    "Error": { "value": "Error" }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": { "type": "string" },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/favorites/relations": {
            "get": {
                "tags": [ "Favorites" ],
                "description": "<h2>Get Favorites List with related items</h2><br><p>Related:</p><p>-> User;</p><p>-> Video</p><br><p>Role required: <strong>Admin</strong></p>",
                "responses": {
                    "200": {
                        "description": "Favorites List with Relations",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Favorites" },
                                "examples": {
                                    "Favorite List ": [
                                        {
                                            "id": 4,
                                            "userId": 3,
                                            "videoId": 2,
                                            "createdAt": "2020-12-04T08:33:54.000Z",
                                            "updatedAt": "2020-12-04T08:33:54.000Z",
                                            "Video": {
                                                "id": 2,
                                                "title": "teste title 2",
                                                "youtubeid": "testeyoutubeid 2",
                                                "status": true,
                                                "createdAt": "2020-12-01T16:14:47.000Z",
                                                "updatedAt": "2020-12-01T16:14:47.000Z"
                                            },
                                            "User": {
                                                "id": 3,
                                                "username": "registered",
                                                "email": "registered@mail.pt",
                                                "password": "$2b$10$yUhKHY3B/lgSEiZjVpz/4OtRxXwY40HtEaW8kCqsdvizMwGgKp9oi",
                                                "status": true,
                                                "role_id": 1,
                                                "createdAt": "2020-11-28T23:43:35.000Z",
                                                "updatedAt": "2020-12-01T09:39:59.000Z"
                                            }
                                        },
                                        {
                                            "id": 6,
                                            "userId": 3,
                                            "videoId": 3,
                                            "createdAt": "2020-12-04T16:23:50.000Z",
                                            "updatedAt": "2020-12-04T16:23:50.000Z",
                                            "Video": {
                                                "id": 3,
                                                "title": "teste title 3",
                                                "youtubeid": "testeyoutubeid 3",
                                                "status": true,
                                                "createdAt": "2020-12-01T16:51:13.000Z",
                                                "updatedAt": "2020-12-01T16:51:13.000Z"
                                            },
                                            "User": {
                                                "id": 3,
                                                "username": "registered",
                                                "email": "registered@mail.pt",
                                                "password": "$2b$10$yUhKHY3B/lgSEiZjVpz/4OtRxXwY40HtEaW8kCqsdvizMwGgKp9oi",
                                                "status": true,
                                                "role_id": 1,
                                                "createdAt": "2020-11-28T23:43:35.000Z",
                                                "updatedAt": "2020-12-01T09:39:59.000Z"
                                            }
                                        }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/favorites/{id}": {
            "get": {
                "tags": [ "Favorites" ],
                "description": "<h2>Get Favorite Info by Id</h2><br><p>Role required <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favorite Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite Info",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Favorites" },
                                "examples": {
                                    "Favorite": {
                                        "id": 2,
                                        "userId": 3,
                                        "videoId": 1,
                                        "createdAt": "2020-12-02T17:42:16.000Z",
                                        "updatedAt": "2020-12-02T17:42:16.000Z"
                                    }
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "put": {
                "tags": [ "Favorites" ],
                "description": "<h2>Update Favorite</h2><br><p>Role required: <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favorite Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Favorites" },
                            "examples": {
                                "Topic": {
                                    "value": {
                                        "topic": "Updated Topic",
                                        "description": "Updated Topic description",
                                        "hexcolor": "#F1C40F"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Topic",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Topics" }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": { 
                                        "value": "Error" 
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "description": "<h2>Delete Favorite</h2><br><p>Role required: <strong>Super Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favorite Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "Favorite Deleted"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/favorites/{id}/relations": {
            "get": {
                "tags": [ "Favorites" ],
                "description": "<h2>Get Favorite Info by Id with relations</h2><br><p>Relations:</p><p>-> User;</p><p>-> Video</p><br><p>Role required <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Favorite Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite Info with relations",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Favorites" },
                                "examples": {
                                    "Favorite": {
                                        "id": 4,
                                        "userId": 3,
                                        "videoId": 2,
                                        "createdAt": "2020-12-04T08:33:54.000Z",
                                        "updatedAt": "2020-12-04T08:33:54.000Z",
                                        "Video": {
                                            "id": 2,
                                            "title": "teste title 2",
                                            "youtubeid": "testeyoutubeid 2",
                                            "status": true,
                                            "createdAt": "2020-12-01T16:14:47.000Z",
                                            "updatedAt": "2020-12-01T16:14:47.000Z"
                                        },
                                        "User": {
                                            "id": 3,
                                            "username": "registered",
                                            "email": "registered@mail.pt",
                                            "password": "$2b$10$yUhKHY3B/lgSEiZjVpz/4OtRxXwY40HtEaW8kCqsdvizMwGgKp9oi",
                                            "status": true,
                                            "role_id": 1,
                                            "createdAt": "2020-11-28T23:43:35.000Z",
                                            "updatedAt": "2020-12-01T09:39:59.000Z"
                                        }
                                    }
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
...
    }
}